Tundra logging API. Existing code modified to use that instead of using Urho log functions directly.

+++ src/TundraCore/Framework/CoreStringUtils.cpp
@@ -73,7 +73,7 @@ String PadString(String str, int pad)
     int orig = pad;
     int to = Abs(pad);
     pad = Abs(pad);
-    while (pad > 0 && str.Length() < to)
+    while (pad > 0 && (int)str.Length() < to)
     {
         if (orig < 0)
             str = " " + str;
